backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
psycopg2-binary
jinja2
